+++ packages/helux-core/src/types/api.d.ts
@@ -1,6 +1,6 @@
 /*
 |------------------------------------------------------------------------------------------------
-| helux-core@3.6.4
+| helux-core@3.6.6
 | A state library core that integrates atom, signal, collection dep, derive and watch,
 | it supports all react like frameworks ( including react 18 ).
 |------------------------------------------------------------------------------------------------
@@ -63,19 +63,18 @@ import type {
   WatchOptionsType,
 } from './base';
 
-export declare const VER: '3.6.4';
-
-export declare const LIMU_VER: string;
-
-export declare const EVENT_NAME: {
-  ON_DATA_CHANGED: 'ON_DATA_CHANGED';
-  ON_SHARE_CREATED: 'ON_SHARE_CREATED';
-};
-
-export declare const RECORD_LOADING: {
-  NONE: NoRecord;
-  PRIVATE: 'private';
-  GLOBAL: 'global';
+export declare const cst: {
+  VER: '3.6.6';
+  LIMU_VER: string;
+  EVENT_NAME: {
+    ON_DATA_CHANGED: 'ON_DATA_CHANGED';
+    ON_SHARE_CREATED: 'ON_SHARE_CREATED';
+  };
+  RECORD_LOADING: {
+    NONE: NoRecord;
+    PRIVATE: 'private';
+    GLOBAL: 'global';
+  };
 };
 
 /**