+++ packages/helux/src/index.ts
@@ -82,8 +82,5 @@ export const {
   getAtom,
   addMiddleware,
   addPlugin,
-  EVENT_NAME,
-  RECORD_LOADING,
-  VER,
-  LIMU_VER,
+  cst,
 } = api;